backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
bcrypt==4.2.0
python-dotenv==1.0.1
pytest==8.3.3
hypercorn==0.17.3